cqrs config
