+++ Core/Core.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 	
   <ItemGroup>
-	<PackageReference Include="MediatR" Version="12.1.1" />
+	<PackageReference Include="MediatR" Version="11.1.0" />
   </ItemGroup>	
 
 </Project>
