--- Web.API/Web.API/Program.cs
@@ -1,48 +1,0 @@
-using Infrastructure.Data.Data;
-using Microsoft.EntityFrameworkCore;
-
-namespace Web.API
-{
-	public class Program
-	{
-		public static void Main(string[] args)
-		{
-			var builder = WebApplication.CreateBuilder(args);
-
-			builder.Services.AddDbContext<BookShopDBContext>(options =>
-						options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
-
-			// Add services to the container.
-			builder.Services.AddAuthorization();
-
-
-			var app = builder.Build();
-
-			// Configure the HTTP request pipeline.
-
-			app.UseHttpsRedirection();
-
-			app.UseAuthorization();
-
-			var summaries = new[]
-			{
-			"Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-		};
-
-			app.MapGet("/weatherforecast", (HttpContext httpContext) =>
-			{
-				var forecast = Enumerable.Range(1, 5).Select(index =>
-					new WeatherForecast
-					{
-						Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-						TemperatureC = Random.Shared.Next(-20, 55),
-						Summary = summaries[Random.Shared.Next(summaries.Length)]
-					})
-					.ToArray();
-				return forecast;
-			});
-
-			app.Run();
-		}
-	}
-}
